requirements.txt
Flask
PyPDF2
pytesseract
Pillow
pdfminer.six
ipywidgets
pymupdf
nltk
nameparser
requests
